requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
cachetools==5.2.1
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.89.1
fastapi-crudrouter==0.8.5
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.10.1
google-cloud-speech==2.17.3
h11==0.14.0
httplib2==0.21.0
idna==3.4
multidict==6.0.4
openai==0.26.1
protobuf==4.21.12
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pydub==0.25.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
starlette==0.22.0
tqdm==4.64.1
twilio==7.16.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
transitions==0.9.0
langchain==0.0.74
faiss-gpu==1.7.2
python-telegram-bot==20.0
asyncpg==0.27.0